location change + bug fixes
