+++ lightautoml/addons/interpretation/__init__.py
@@ -1,6 +1,5 @@
 from .l2x import L2XTextExplainer
 from .lime import LimeTextExplainer
-from .sswarm import SSWARM
 
 
 __all__ = ["LimeTextExplainer", "L2XTextExplainer", "SSWARM"]
