+++ lightautoml/addons/tabular_interpretation/__init__.py
@@ -1,0 +1,4 @@
+from .sswarm import SSWARM
+
+
+__all__ = ["SSWARM"]
